ai-services/interview_analysis/requirements.txt
# AI Voice Interview Agent - Free Open Source Dependencies

# Core AI Framework
google-generativeai>=0.3.0

# Database
pymongo>=4.6.0

# Voice Processing (Free & Open Source)
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13

# Audio Processing
pydub>=0.25.1

# Document Processing (Free)
pdfplumber>=0.9.0
docx2txt>=0.8

# Natural Language Processing (Free)
spacy>=3.7.0

# Data Management
python-dotenv>=1.0.0
pandas>=2.0.0

# Web Framework for UI (Optional - Free)
streamlit>=1.28.0
Flask>=2.3.0

# Utilities
requests>=2.31.0
